+++ changelog.md
@@ -1,8 +1,16 @@
-#### in progress:
+## in progress:
 
+#### Features
+
+- New spring friendship commands added
 - 
 
-#### v1.1.1
+#### Bugfixes
+
+- Relationships report now correctly combines Block/Avoid
+- Relationships report correctly reports on Ignore counts
+
+## v1.1.1
 
 * Report forms now include the option to ignore players along with buttons to avoid or block them
 * Ignoring is now separate from avoiding or blocking
@@ -10,6 +18,6 @@
 * Contributors can set their own flag
 * Moderation action updates/deletions now propagate to the discord
 
-#### v1.1.0
+## v1.1.0
 
 - I didn't bother with a changelog because I was lazy and kept putting it off